Modificari in functia de scalare. Modificare tip de data factor in float pentru valori subunitare (scalare negativa). Adaugarea unui mod diferide a gasi x optim prelucrarii de cadre intr-o scalare cu factorul <1

+++ Core/Inc/graphics.h
@@ -73,7 +73,7 @@ void draw_rectangle(uint16_t x0, uint16_t y0, uint16_t x1, uint16_t y1, uint16_t
 void draw_entity(ENTITY *entity, char *filePathName);
 void translation_entity(ENTITY *const restrict entity, int16_t x, int16_t y, bool TurnOnStep);//, uint16_t color);
 void translation_test(ENTITY *entity, uint8_t step, uint16_t delay);
-void scaling_entity(ENTITY *entity, const uint8_t factor, char *filePathName);
+void scaling_entity(ENTITY *entity, const float factor, char *filePathName);
 
 #endif /* INC_GRAPHICS_H_ */
